+++ src/app/services/characters.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import { Character } from '../model/character';
+import { GiftData } from '../model/giftData';
 
 @Injectable({
   providedIn: 'root'
@@ -12,6 +13,13 @@ export class CharactersService {
 
   public character: BehaviorSubject<any> = new BehaviorSubject(null);
   public character$ = this.character.asObservable();
+
+  public giftData: BehaviorSubject<any> = new BehaviorSubject(null);
+  public giftData$ =this.giftData.asObservable();
+
+  public setGiftData(data: GiftData): void {
+    this.giftData.next(data);
+  }
 
   public setCharacters(characters: Character[]): void {
     this.characters.next(characters);